+++ activitysampling-application/src/main/java/de/muspellheim/activitysampling/application/App.java
@@ -1,5 +1,6 @@
 package de.muspellheim.activitysampling.application;
 
+import de.muspellheim.activitysampling.application.activitysampling.*;
 import javafx.application.*;
 import javafx.stage.*;