+++ activitysampling-application/src/main/java/de/muspellheim/activitysampling/application/ViewModels.java
@@ -1,24 +1,26 @@
 package de.muspellheim.activitysampling.application;
 
+import de.muspellheim.activitysampling.application.activitysampling.*;
+import de.muspellheim.activitysampling.application.timesheet.*;
 import de.muspellheim.activitysampling.domain.*;
 import de.muspellheim.activitysampling.infrastructure.*;
 import java.nio.file.*;
 import java.util.function.*;
 
-class ViewModels {
+public class ViewModels {
   private ViewModels() {
     // do not instantiate static class
   }
 
-  static ActivitySamplingViewModel newActivitySampling(
+  public static ActivitySamplingViewModel newActivitySampling(
       Runnable onCountdownElapsed, Consumer<String> onError) {
     var activitiesService = getActivitiesService();
     return new ActivitySamplingViewModel(activitiesService, onCountdownElapsed, onError);
   }
 
-  static TimesheetViewModel newTimesheet() {
+  public static TimesheetViewModel newTimesheet(Consumer<String> onError) {
     var activitiesService = getActivitiesService();
-    return new TimesheetViewModel(activitiesService);
+    return new TimesheetViewModel(activitiesService, onError);
   }
 
   private static ActivitiesServiceImpl getActivitiesService() {