+++ activitysampling-application/src/main/java/de/muspellheim/activitysampling/application/shared/ErrorView.java
@@ -1,0 +1,17 @@
+package de.muspellheim.activitysampling.application.shared;
+
+import javafx.scene.control.*;
+
+public class ErrorView {
+  private ErrorView() {
+    // Do not instantiate static class.
+  }
+
+  public static void handleError(String message) {
+    var alert = new Alert(Alert.AlertType.ERROR);
+    alert.setTitle("Error");
+    alert.setHeaderText("An unexpected error occurred.");
+    alert.setContentText(message);
+    alert.show();
+  }
+}