+++ activitysampling-application/src/main/java/de/muspellheim/activitysampling/application/timesheet/TimesheetItem.java
@@ -1,9 +1,9 @@
-package de.muspellheim.activitysampling.application;
+package de.muspellheim.activitysampling.application.timesheet;
 
 import java.util.*;
 
-record TimesheetItem(String date, String activity, String duration) {
-  TimesheetItem {
+public record TimesheetItem(String date, String activity, String duration) {
+  public TimesheetItem {
     Objects.requireNonNull(date, "date");
     Objects.requireNonNull(activity, "activity");
     Objects.requireNonNull(duration, "duration");