+++ activitysampling-application/src/main/java/module-info.java
@@ -9,4 +9,10 @@ module de.muspellheim.activitysampling.application {
   opens de.muspellheim.activitysampling.application to
       javafx.fxml,
       javafx.graphics;
+  opens de.muspellheim.activitysampling.application.activitysampling to
+      javafx.fxml,
+      javafx.graphics;
+  opens de.muspellheim.activitysampling.application.timesheet to
+      javafx.fxml,
+      javafx.graphics;
 }