+++ activitysampling-domain/src/main/java/de/muspellheim/activitysampling/domain/TimesheetByDayProjection.java
@@ -3,13 +3,13 @@ package de.muspellheim.activitysampling.domain;
 import java.time.*;
 import java.util.*;
 
-class TimesheetProjection {
+class TimesheetByDayProjection {
   private final LocalDate from;
   private final LocalDate to;
 
   private final SortedMap<LocalDate, SortedMap<String, Duration>> entries = new TreeMap<>();
 
-  TimesheetProjection(LocalDate from, LocalDate to) {
+  TimesheetByDayProjection(LocalDate from, LocalDate to) {
     this.from = from;
     this.to = to;
   }
@@ -34,19 +34,23 @@ class TimesheetProjection {
     activities.put(event.description(), duration);
   }
 
-  List<WorkingDay> get() {
-    return null;
-    /*
+  List<TimesheetByDay> get() {
     return entries.entrySet().stream()
         .map(
             d ->
-                new WorkingDay(
-                    d.getKey(),
-                    d.getValue().entrySet().stream()
-                        .map(e -> new TimesheetEntry(d.getKey(), e.getKey(), e.getValue()))
-                        .toList()))
+                new TimesheetByDay(
+                    d.getKey(), getTimesheetEntries(d), getSubtotal(d.getValue().values())))
         .toList();
+  }
 
-    */
+  private static List<TimesheetEntry> getTimesheetEntries(
+      Map.Entry<LocalDate, SortedMap<String, Duration>> d) {
+    return d.getValue().entrySet().stream()
+        .map(e -> new TimesheetEntry(e.getKey(), e.getValue()))
+        .toList();
+  }
+
+  private static Duration getSubtotal(Collection<Duration> durations) {
+    return durations.stream().reduce(Duration::plus).orElse(Duration.ZERO);
   }
 }