+++ activitysampling-domain/src/main/java/de/muspellheim/activitysampling/domain/TimesheetEntry.java
@@ -2,4 +2,4 @@ package de.muspellheim.activitysampling.domain;
 
 import java.time.*;
 
-public record TimesheetEntry(LocalDate date, String activity, Duration duration) {}
+public record TimesheetEntry(String activity, Duration duration) {}