+++ activitysampling-domain/src/test/java/de/muspellheim/activitysampling/domain/ActivitiesServiceTests.java
@@ -203,37 +203,31 @@ class ActivitiesServiceTests {
                     Instant.parse("2022-11-19T07:00:00Z"), Duration.ofMinutes(20), "A2")));
 
     var timesheet = sut.createTimesheet(LocalDate.of(2022, 11, 14), LocalDate.of(2022, 11, 18));
-    /*
+
     assertEquals(
         new Timesheet(
             List.of(
-                // FIXME missing subtotal per day
-                new WorkingDay(
+                new TimesheetByDay(
                     LocalDate.of(2022, 11, 14),
-                    List.of(
-                        new TimesheetEntry(
-                            LocalDate.of(2022, 11, 14), "A1", Duration.ofMinutes(20)))),
-                new WorkingDay(
+                    List.of(new TimesheetEntry("A1", Duration.ofMinutes(20))),
+                    Duration.ofMinutes(20)),
+                new TimesheetByDay(
                     LocalDate.of(2022, 11, 15),
                     List.of(
-                        new TimesheetEntry(
-                            LocalDate.of(2022, 11, 15), "A1", Duration.ofMinutes(20)),
-                        new TimesheetEntry(
-                            LocalDate.of(2022, 11, 15), "A2", Duration.ofMinutes(40)))),
-                new WorkingDay(
+                        new TimesheetEntry("A1", Duration.ofMinutes(20)),
+                        new TimesheetEntry("A2", Duration.ofMinutes(40))),
+                    Duration.ofMinutes(60)),
+                new TimesheetByDay(
                     LocalDate.of(2022, 11, 17),
                     List.of(
-                        new TimesheetEntry(
-                            LocalDate.of(2022, 11, 17), "A1", Duration.ofMinutes(40)),
-                        new TimesheetEntry(
-                            LocalDate.of(2022, 11, 17), "A2", Duration.ofMinutes(20)))),
-                new WorkingDay(
+                        new TimesheetEntry("A1", Duration.ofMinutes(40)),
+                        new TimesheetEntry("A2", Duration.ofMinutes(20))),
+                    Duration.ofMinutes(60)),
+                new TimesheetByDay(
                     LocalDate.of(2022, 11, 18),
-                    List.of(
-                        new TimesheetEntry(
-                            LocalDate.of(2022, 11, 18), "A2", Duration.ofMinutes(20))))),
+                    List.of(new TimesheetEntry("A2", Duration.ofMinutes(20))),
+                    Duration.ofMinutes(20))),
             Duration.ofMinutes(160)),
         timesheet);
-    */
   }
 }