+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+## Added
+
+- Create timesheet with total hours per activity and day.
+
 ## [1.1.0] - 2022-11-17
 
 ### Added